+++ iosHyperskillApp/iosHyperskillApp/Sources/Modules/StepQuiz/ViewData/StepQuizViewData.swift
@@ -10,6 +10,4 @@ struct StepQuizViewData {
     let quizName: String?
 
     let feedbackHintText: String?
-
-    let stepHasHints: Bool
 }
